Auth and User API refactor

+++ api/controllers/user.controller.js
@@ -2,17 +2,35 @@ import User from "../models/user.model.js";
 import createError from "../utils/createError.js";
 
 export const deleteUser = async (req, res, next) => {
-  const user = await User.findById(req.params.id);
+  try {
+    const user = await User.findById(req.params.id);
+    
+    if (!user) {
+      return next(createError(404, "User not found."));
+    }
 
-  if (req.userId !== user._id.toString()) {
-    return next(createError(403, "You can delete only your account!"));
+    if (req.userId !== user._id.toString()) {
+      return next(createError(403, "You can delete only your account!"));
+    }
+
+    await User.findByIdAndDelete(req.params.id);
+    res.status(200).json({ message: "User has been deleted." });
+  } catch (error) {
+    next(error);
   }
-  await User.findByIdAndDelete(req.params.id);
-  res.status(200).send("User has been deleted.");
 };
 
 export const getUser = async (req, res, next) => {
-  const user = await User.findById(req.params.id);
+  try {
+    const user = await User.findById(req.params.id);
 
-  res.status(200).send(user);
+    if (!user) {
+      return next(createError(404, "User not found."));
+    }
+
+    const { password, ...userInfo } = user._doc;
+    res.status(200).json(userInfo);
+  } catch (error) {
+    next(error);
+  }
 };
